deployment/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1
Werkzeug==2.3.7
flask-socketio==5.3.2
python-socketio==5.8.0
eventlet==0.33.3
requests==2.31.0
APScheduler==3.10.4
